new fake global function and free at heredoc child qith signal

+++ src/free_loop.c
@@ -22,7 +22,7 @@ void	free_token_lst(t_token **token_lst)
 	{
 		tmp = (*token_lst)->next;
 		if ((*token_lst)->value)
-      free((*token_lst)->value);
+			free((*token_lst)->value);
 		if ((*token_lst)->expand_var)
 			free((*token_lst)->expand_var);
 		free(*token_lst);
@@ -31,14 +31,15 @@ void	free_token_lst(t_token **token_lst)
 	*token_lst = NULL;
 }
 
-static void	free_redirections(t_redirections *redir)
+static void	free_redirections(t_shell *shell, t_redirections *redir)
 {
 	t_redirections	*tmp;
 
 	while (redir)
 	{
 		tmp = redir->next;
-		if (ft_strncmp(redir->filename, "/tmp/.hdoc_tmp_", 15) == 0)
+		if (!shell->hdoc_control && \
+			ft_strncmp(redir->filename, "/tmp/.hdoc_tmp_", 15) == 0)
 			unlink(redir->filename);
 		free(redir->filename);
 		free(redir);
@@ -46,7 +47,7 @@ static void	free_redirections(t_redirections *redir)
 	}
 }
 
-void	free_command_list(t_command *head)
+void	free_command_list(t_shell *shell, t_command *head)
 {
 	t_command	*tmp;
 	int			i;
@@ -64,19 +65,20 @@ void	free_command_list(t_command *head)
 			free(head->args);
 		}
 		if (head->redirs)
-			free_redirections(head->redirs);
+			free_redirections(shell, head->redirs);
 		free(head);
 		head = tmp;
 	}
 }
 
-void	free_loop(t_token **token_lst, t_command **cmd)
+// void	free_loop(t_token **token_lst, t_command **cmd, int hdoc_control)
+void	free_loop(t_shell *shell)
 {
-	if (*token_lst)
-		free_token_lst(token_lst);
-	if (*cmd)
+	if (shell->token_list)
+		free_token_lst(&shell->token_list);
+	if (shell->cmd)
 	{
-		free_command_list(*cmd);
-		*cmd = NULL;
+		free_command_list(shell, shell->cmd);
+		shell->cmd = NULL;
 	}
 }
